action-system/requirements.txt
Brotli==1.0.9
click==8.0.1
confluent-kafka==1.7.0
cycler==0.10.0
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
plotly==5.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
tenacity==8.0.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
